server/requirements.txt
Flask==2.1.1
psycopg2-binary==2.9.3
Flask-CORS==3.0.10
Werkzeug==2.0.3